Beispieldaten/requirements.txt
pandas
streamlit
plotly-express
numpy
xlsxwriter